First row is a visible variant
Previous to this change the first row was treated as the master variant
and as a result was not visible. By not setting the product sku, the
master sku is empty string. Then later on down the line, the row is
treated as a new variant and is added to product#variants.